python3: fix _acceptable_args del iteration
dict.items returns a ItemsView iterator in python3; while iterating, if
the size of the underlying dictionary changes, RuntimeError is raised.

Iterating through the entire ItemsView before performing the item
deletions is one way to handle this in python3, and is equivalent to
what happens in python2's dict.items.